feat: add SendCmioResponse
