+++ shasum-mock
@@ -2,7 +2,7 @@ c114d9e59014339c29374f0085a266fb820449f123007e7c89f2f9ec680e5de0  src/AccessLogs
 b0bb91ae378db1c938e73eccde20c539da01b8556165835839fd1794ca573404  src/Buffer.sol
 c2e9abd5557092fb2c7a3b6eb3d55933613c202b8ce54bba2e8cbd2d276278a1  src/Memory.sol
 06df1c5a0a6094fdc3f55240a42b1991290eb672fae6833e9f2b4d4bfe2a5b88  src/MetaStep.sol
-d858085617f1e8037e06420d9569e2196b72a8bd0dd217aefbb8bb738b44a8fd  src/UArchCompat.sol
-1152e2e0f1fdc8befe075251f5e69dcc55ec2666a287fc5002bfa021b2aaea98  src/UArchConstants.sol
+d858085617f1e8037e06420d9569e2196b72a8bd0dd217aefbb8bb738b44a8fd  src/EmulatorCompat.sol
+1152e2e0f1fdc8befe075251f5e69dcc55ec2666a287fc5002bfa021b2aaea98  src/EmulatorConstants.sol
 c250ea17a7a471716148a60528b2cd01bd82dffd1c3c57aa0a7e99527fe44527  src/UArchReset.sol
 40ab03ff44235b8ee5490314db112120e1ff99b347a5faf6a00f812e2497568d  src/UArchStep.sol
