+++ shasum-prod
@@ -2,7 +2,7 @@
 b0bb91ae378db1c938e73eccde20c539da01b8556165835839fd1794ca573404  src/Buffer.sol
 d5de348b450c7cb5237a836d8e6e22cbfb4649f5892fac30f054efabdd7fc0a5  src/Memory.sol
 06df1c5a0a6094fdc3f55240a42b1991290eb672fae6833e9f2b4d4bfe2a5b88  src/MetaStep.sol
-0cdad8286752fbecf920db8095a764802ed88236efe7d179acdd88e018291b33  src/UArchCompat.sol
-60ab17c719b0cb58047bfa5a83e390b30b32f5be00243890f2d321635bc5359a  src/UArchConstants.sol
+0cdad8286752fbecf920db8095a764802ed88236efe7d179acdd88e018291b33  src/EmulatorCompat.sol
+60ab17c719b0cb58047bfa5a83e390b30b32f5be00243890f2d321635bc5359a  src/EmulatorConstants.sol
 c250ea17a7a471716148a60528b2cd01bd82dffd1c3c57aa0a7e99527fe44527  src/UArchReset.sol
 40ab03ff44235b8ee5490314db112120e1ff99b347a5faf6a00f812e2497568d  src/UArchStep.sol
