+++ templates/SendCmioResponse.sol.template
@@ -14,16 +14,19 @@
 // limitations under the License.
 //
 
-/// @title UArchConstants
-/// @notice Contains constants for micro-architecture
+/// @title SendCmioResponse
+/// @notice Reset microarchitecture to pristine state
 //:#include macro.pp
-/// DEV_COMMENT(templates/UArchConstants.sol.template)
+/// DEV_COMMENT(templates/UArchReset.sol.template)
 
 pragma solidity ^0.8.0;
 
-library UArchConstants {
+import "./EmulatorCompat.sol";
+
+library SendCmioResponse {
+    using Memory for uint64;
+    using AccessLogs for AccessLogs.Context;
+    
     // START OF AUTO-GENERATED CODE
     // END OF AUTO-GENERATED CODE
-
-    uint64 constant LOG2_CYCLES_TO_RESET = 10;
 }
